+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 ### Changed
 
 - Improved logging.
+- `Struct\PropertySetStruct::exportItems()` now groups all custom properties under the "properties"
+  key.
 
 
 ## 0.2 - 2016-08-23